COM-3690: sophie's review
